+++ app/models/trade.rb
@@ -45,7 +45,7 @@ class Trade < ApplicationRecord
   end
 
   def trade_style_win_rate(trade_style_id)
-    return "0%" if win_number_of_trades === 0
+    return "0%" if win_number_of_trades.count === 0
     trades = win_number_of_trades.where("trade_style_id" => trade_style_id).count
     "#{(trades / user_trades.count.to_f * 100).floor(1)}%"
   end
